+++ codegen/src/rust.rs
@@ -961,6 +961,7 @@ fn type_name(ty: &ast::TypeName) -> String {
         ast::TypeNameKind::Receiver(ty) => {
             format!("aldrin::UnboundReceiver<{}>", type_name(ty))
         }
+        ast::TypeNameKind::Lifetime => "aldrin::LifetimeId".to_owned(),
         ast::TypeNameKind::Extern(m, ty) => format!("super::{}::{}", m.value(), ty.value()),
         ast::TypeNameKind::Intern(ty) => ty.value().to_owned(),
     }
@@ -1003,6 +1004,7 @@ fn call_type_name(ty: &ast::TypeName) -> String {
         ast::TypeNameKind::Receiver(ty) => {
             format!("aldrin::UnboundReceiver<{}>", type_name(ty))
         }
+        ast::TypeNameKind::Lifetime => "aldrin::LifetimeId".to_owned(),
         ast::TypeNameKind::Extern(m, ty) => format!("&super::{}::{}", m.value(), ty.value()),
         ast::TypeNameKind::Intern(ty) => format!("&{}", ty.value()),
     }
@@ -1045,6 +1047,7 @@ fn sender_type_name(ty: &ast::TypeName) -> String {
         ast::TypeNameKind::Receiver(ty) => {
             format!("aldrin::UnboundReceiver<{}>", type_name(ty))
         }
+        ast::TypeNameKind::Lifetime => "aldrin::LifetimeId".to_owned(),
         ast::TypeNameKind::Extern(m, ty) => format!("super::{}::{}", m.value(), ty.value()),
         ast::TypeNameKind::Intern(ty) => ty.value().to_owned(),
     }