+++ codegen/CHANGELOG.md
@@ -1,6 +1,14 @@
+# Unreleased
+
+## Added
+
+- Support `lifetime` built-in type. It resolves to `aldrin::LifetimeId` in all cases.
+
+
 # 0.2.0 (November 27th, 2023)
 
 - Bump for Aldrin 0.2.0.
+
 
 # 0.1.0 (November 24th, 2023)