+++ parser/CHANGELOG.md
@@ -1,6 +1,14 @@
+# Unreleased
+
+## Added
+
+- Add built-in type `lifetime`.
+
+
 # 0.2.0 (November 27th, 2023)
 
 - Bump for Aldrin 0.2.0.
+
 
 # 0.1.0 (November 24th, 2023)